Redesign user menu at navbar
