+++ src/components/NavBar/NavBar.js
@@ -1,8 +1,5 @@
 import React, { useEffect, useState, lazy, Suspense } from "react";
-import { AiOutlineUser } from "react-icons/ai";
-import { FiEdit3 } from 'react-icons/fi';
-import { GrAchievement } from 'react-icons/gr';
-import { MdLogout, MdLogin, MdDelete } from "react-icons/md";
+import { MdLogin } from "react-icons/md";
 import { Link } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 
@@ -19,9 +16,7 @@ import {
 import Loading from "../../utils/Loading/Loading";
 
 const DeletePopup = lazy(() => import("../../utils/DeletePopup/DeletePopup"));
-const TodoList = lazy(() => import('../../icons/list/TodoList'));
-const Menu = lazy(() => import("../../utils/Menu/Menu"));
-const MenuItem = lazy(() => import("../../utils/Menu/MenuItem"));
+const UserMenu = lazy(() => import('./UserMenu.js'))
 
 const NavBar = ({ setMessage }) => {
   const [openDelete, setOpenDelete] = useState(false);
@@ -55,28 +50,10 @@ const NavBar = ({ setMessage }) => {
     // eslint-disable-next-line
   }, [user]);
 
-  const logout = () => {
-    dispatch({ type: LOGOUT });
-  };
-
   const handleDeleteUser = () => {
     setOpenDelete(false);
     dispatch(deleteUser(setMessage));
   }
-
-  const getUserName = () => {
-    if (user.name.length <= 18) {
-      return user.name;
-    } else if (user.name.split(' ').length > 1) {
-      return user.name.split(' ')[0] + " " + user.name.split(' ')[1][0] + ".";
-    } else {
-      return user.name.slice(0, 15) + '...';
-    }
-  }
-
-  const toggleDelete = () => {
-    setOpenDelete(true);
-  };
 
   return (
     <>
@@ -126,66 +103,7 @@ const NavBar = ({ setMessage }) => {
                   />
                 </>
               }>
-                <Menu MainButton={
-                  <button
-                    aria-label="user button"
-                    className="user-menu"
-                  >
-                    <div className={`user-logo ${user?.avatar && 'img'}`}>
-                      {user?.avatar ? (
-                        <img src={user?.avatar} alt="avatar" width="" height="" />
-                      ) : (
-                        <AiOutlineUser />
-                      )}
-                    </div>
-                    <span className="name">{getUserName()}</span>
-                  </button>
-                }>
-                  <MenuItem
-                    component={Link}
-                    to="/user/achievements"
-                    aria-label="user button in menu"
-                  >
-                    <GrAchievement />
-                    <span style={{ marginLeft: 10 }}>
-                      achievements
-                    </span>
-                  </MenuItem>
-                  <MenuItem
-                    component={Link}
-                    to="/templates/you"
-                    aria-label="user button in menu">
-                    <TodoList />
-                    <span style={{ marginLeft: 10 }}>
-                      todos templates
-                    </span>
-                  </MenuItem>
-                  <MenuItem
-                    component={Link}
-                    to="/user/edit"
-                    aria-label="user button in menu">
-                    <FiEdit3 />
-                    <span style={{ marginLeft: 10 }}>
-                      edit account
-                    </span>
-                  </MenuItem>
-                  <MenuItem
-                    onClick={logout}
-                    aria-label="logout button in menu">
-                    <MdLogout />
-                    <span style={{ marginLeft: 10 }}>
-                      Logout
-                    </span>
-                  </MenuItem>
-                  <MenuItem
-                    aria-label="delete account button"
-                    onClick={toggleDelete}
-                    style={{ color: "red" }}
-                  >
-                    <MdDelete />
-                    <span style={{ marginLeft: 10 }}>Delete Account</span>
-                  </MenuItem>
-                </Menu>
+                <UserMenu setOpenDelete={setOpenDelete} />
               </Suspense>
             </>
           ) : (
