+++ src/components/NavBar/style.css
@@ -41,6 +41,11 @@
   font-size: 20px;
   color: #fff;
   background: #dadada99;
+
+  width: 45px;
+  height: 45px;
+  border-radius: 50%;
+  font-size: 22px;
 }
 
 .nav-bar .user-menu:hover {
@@ -63,14 +68,53 @@
   border: 2px solid;
 }
 
-.nav-bar .user-menu .user-logo.img {
+.nav-bar .user-menu .img {
   padding: 0px;
 }
 
-.nav-bar .user-menu .user-logo img {
-  width: 40px;
-  height: 40px;
-  border-radius: 9px;
+.nav-bar .user-menu img {
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
+}
+
+.nav-bar .menu .menu-avatar {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.nav-bar .menu .menu-avatar img,
+.nav-bar .menu .menu-avatar {
+  width: 100px;
+  height: 100px;
+  border-radius: 50%;
+}
+
+.nav-bar .menu .user-details {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  padding: 16px;
+  font-size: 20px;
+}
+
+.nav-bar .menu .user-details .user-email {
+  font-size: 16px;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  max-width: 200px;
+}
+
+.nav-bar .menu .user-details .menu-avatar {
+  font-size: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100px;
+  height: 100px;
+  color: #fff;
 }
 
 .nav-bar h1 {
